NXTornadoWServer/requirements.txt
tornado
nxt-python
pybluez